+++ src/main/java/chess/dao/PieceDaoImpl.java
@@ -1,0 +1,30 @@
+package chess.dao;
+
+import chess.dto.PieceResponse;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+
+public class PieceDaoImpl implements PieceDao {
+
+    @Override
+    public Long save(final PieceResponse pieceResponse, final Long gameId, final Connection connection) {
+        String query = "INSERT INTO piece (game_id, x, y, color, type) VALUES (?, ?, ?, ?, ?)";
+        try (PreparedStatement preparedStatement = connection.prepareStatement(query,
+                Statement.RETURN_GENERATED_KEYS)) {
+            preparedStatement.setLong(1, gameId);
+            preparedStatement.setInt(2, pieceResponse.rankIndex());
+            preparedStatement.setInt(3, pieceResponse.fileIndex());
+            preparedStatement.setString(4, pieceResponse.color().name());
+            preparedStatement.setString(5, pieceResponse.type().name());
+            preparedStatement.executeUpdate();
+            ResultSet resultSet = preparedStatement.getGeneratedKeys();
+            resultSet.next();
+            return resultSet.getLong(1);
+        } catch (SQLException exception) {
+            throw new RuntimeException(exception);
+        }
+    }
+}
