--- src/main/java/chess/dao/PieceRepository.java
@@ -1,9 +1,0 @@
-package chess.dao;
-
-import chess.dto.PieceResponse;
-import java.sql.Connection;
-
-public interface PieceRepository {
-
-    Long save(PieceResponse pieceResponse, Long gameId, Connection connection);
-}
