+++ src/test/java/chess/dao/PieceDaoFake.java
@@ -5,19 +5,19 @@ import java.sql.Connection;
 import java.util.HashMap;
 import java.util.Map;
 
-public class PieceDaoFake implements PieceRepository {
-    private final Map<Long, PieceResponse> pieceRepository;
+public class PieceDaoFake implements PieceDao {
+    private final Map<Long, PieceResponse> pieces;
     private Long pieceId;
 
     public PieceDaoFake() {
-        this.pieceRepository = new HashMap<>();
+        this.pieces = new HashMap<>();
         this.pieceId = 0L;
     }
 
     @Override
     public Long save(final PieceResponse pieceResponse, final Long gameId, final Connection connection) {
         pieceId++;
-        pieceRepository.put(gameId, pieceResponse);
+        pieces.put(gameId, pieceResponse);
         return pieceId;
     }
 }
